cli: refactor create factories to use common prompts
